added patch from Kenny to identify if the distance sensor is used or not, depending on this the speed is calculated when lat/lon is the same, as it can be in tunnels
git-svn-id: https://mytourbook.svn.sourceforge.net/svnroot/mytourbook/trunk@1430 5a91af82-4d22-0410-8bcb-a41d7c89129f

+++ net.tourbook/src/net/tourbook/ui/views/tourBook/TVITourBookItem.java
@@ -87,8 +87,8 @@ public abstract class TVITourBookItem extends TreeViewerItem implements ITourIte
 	long					colAvgCadence;
 	long					colAvgTemperature;
 
-	float					colWindSpd;
-	float					colWindDir;
+	int					colWindSpd;
+	int					colWindDir;
 	String					fClouds;
 	int						colRestPulse;
 
@@ -126,8 +126,8 @@ public abstract class TVITourBookItem extends TreeViewerItem implements ITourIte
 		colAvgCadence = result.getLong(startIndex + 12);
 		colAvgTemperature = result.getLong(startIndex + 13);
 
-		colWindDir = result.getLong(startIndex + 14);
-		colWindSpd = result.getLong(startIndex + 15);
+		colWindDir = result.getInt(startIndex + 14);
+		colWindSpd = result.getInt(startIndex + 15);
 		colRestPulse = result.getInt(startIndex + 16);
 
 		colCalories = result.getLong(startIndex + 17);
